% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello name function}
\usage{
hello(n)
}
\arguments{
\item{n}{Name of person to say hello to}
}
\value{
The output from \code{\link{print}}
}
\description{
Hello name function
}
\examples{
hello("John")
}
