% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeblastdb.R
\name{makeblastdb}
\alias{makeblastdb}
\title{Run makeblastdb}
\usage{
makeblastdb(input, out, dbtype)
}
\arguments{
\item{input}{String, path to input fasta file}

\item{out}{String, name of output files}

\item{dbtype}{String, name of database type "nucl" / "prot"}
}
\description{
Run makeblastdb
}
\examples{
makeblastdb()
}
