\name{SC-package}
\alias{SC-package}
\alias{SC}
\docType{package}
\title{
\packageTitle{SC}
}
\description{
\packageDescription{SC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SC}
\packageIndices{SC}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{SC}

Maintainer: \packageMaintainer{SC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
