% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfunction.R
\name{sc}
\alias{sc}
\title{Title}
\usage{
sc(S, lambda1 = 0, lambda2, max_iter = 50, init.x = NULL,
  type = c("fused", "lasso", "l1trend", "HP"), band = NULL,
  ABSTOL = 1e-04)
}
\arguments{
\item{S}{Sample Covariance Matrix}

\item{lambda1}{Threshold value for sparse SC}

\item{lambda2}{Threshold value to control smoothness}

\item{max_iter}{Number of maximum iteration}

\item{init.x}{initial value for vectorized Cholesky factor L. The default is diagonal matrix, with diagonal values corresponding square root of diag(S).}

\item{type}{}

\item{band}{}

\item{ABSTOL}{}
}
\value{
The smoothed Cholesky factor \code{L}
}
\description{
Title
}
