% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scfunction.R
\name{smoothchol}
\alias{smoothchol}
\title{Title}
\usage{
smoothchol(S, lambda1 = 0, lambda2, max_iter = 50, init.x = NULL,
  type = c("fused", "lasso", "l1trend", "HP"), band = NULL,
  ABSTOL = 1e-04)
}
\arguments{
\item{S}{Sample Covariance Matrix}

\item{lambda1}{Threshold value for sparse SC}

\item{lambda2}{Threshold value to control smoothness}

\item{max_iter}{Number of maximum iteration}

\item{init.x}{initial value for vectorized Cholesky factor L. The default is diagonal matrix, with diagonal values corresponding square root of diag(S).}

\item{type}{type of the smoothing penalty}

\item{band}{if specified, algorithm forces the rest of entries zero and iterates only over specifed subdiagonals.}

\item{ABSTOL}{Tolerance for algorithm convergence.}
}
\value{
The smoothed Cholesky factor \code{L}
}
\description{
Title
}
\examples{
set.seed(12)
p <- 50
band <- 5
L_true <- generateL(p = p, band = band)
library(varband)
n = 100
random sample
X <- sample_gen(L = true, n = n)
sample covariance matrix
L_fused = sc(S, lambda1 = 0, lambda2 = 0.2, type = "fused")
S <- crossprod(scale(X, center = TRUE, scale = FALSE)) / n
}
