% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblastn.R
\name{tblastn}
\alias{tblastn}
\title{Run tblastn}
\usage{
tblastn(blast_db, input, evalue)
}
\arguments{
\item{blast_db}{String, path to blast database (nucleotide)}

\item{input}{String, path to input query file (protein fasta)}

\item{evalue}{Integer, evalue}
}
\value{
blast_out
}
\description{
Run tblastn
}
\examples{
tblastn()
}
