% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateLsmooth.R
\name{generateL}
\alias{generateL}
\title{Title}
\usage{
generateL(p, band = NULL, scaled = FALSE, case = c("a", "b", "c",
  "d"), seed = 25, ...)
}
\arguments{
\item{p}{Number of observations}

\item{band}{Number of band}

\item{scaled}{If TRUE the scaled \code{L} is returned}

\item{case}{case as in Dallakyan and Pourahmadi 2019 paper.}

\item{seed}{}

\item{...}{}
}
\value{
Returns Standard Cholesky factor \code{L} and modified Cholesky Factor \code{T}
}
\description{
Title
}
\examples{
generateL(p = 10, band = 4, case = "c")
}
