% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateLsmooth.R
\name{generateL}
\alias{generateL}
\title{Generate a model with specified structure}
\usage{
generateL(p, band = NULL, scaled = FALSE, case = c("a", "b", "c",
  "d"), seed = 25, ...)
}
\arguments{
\item{p}{the dimension of \code{L}}

\item{band}{Number of subdiagonals}

\item{scaled}{If TRUE the diagonals of lower triangular matrix \code{L} are equal 0, otherwise \eqn{log((1:p)/10+2)}}

\item{case}{admits for possible case a,b,c, and d discribed in Dallakyan and Pourahmadi(2019)}

\item{seed}{}

\item{...}{}
}
\value{
Returns Standard Cholesky factor \code{L} and modified Cholesky Factor \code{T}.
}
\description{
Generate a model with structure described below.
}
\details{
By default diagonal elements of \code{L} is of the form \eqn{log((1:p)/10+2}.
}
\examples{
generateL(p = 10, band = 4, case = "c")

}
