% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothcholCV.R
\name{smoothcholCV}
\alias{smoothcholCV}
\title{Title}
\usage{
smoothcholCV(k = 5, X, both.lambda = FALSE, lambda1_seq = NULL,
  lambda2_seq = NULL, max_iter = 50, band = NULL, n_lambda = 60,
  pen.type = c("HP", "fused", "l1trend"), ABSTOL = 0.001,
  stand = FALSE)
}
\arguments{
\item{k}{Folds used in cross-validation. The default is $k = 5$}

\item{X}{A n-by-p sample matrix, each row is an observation of th p-dim random vector.}

\item{both.lambda}{Logical. If TRUE the cross-validation implemented for both lambdas.}

\item{lambda1_seq}{A vector of non-negative tuning parameters for lambda1 to control sparsity.}

\item{lambda2_seq}{A vector of non-negative tuning parameters for lambda1 to control smoothness}

\item{max_iter}{Maximum number of iterations}

\item{band}{Positive number of subdiagonal to be estimated. If not provided, the algorithm iterates over all subdiagonals.}

\item{n_lambda}{If lambda1_seq and lambda2_seq is not provided, create a vector of lambdas with length n_lambda. Default is 60.}

\item{pen.type}{Selects penalty for smoothness.}

\item{ABSTOL}{The tolerence for convergence}

\item{stand}{Logical, if TRUE the data will be standardized.}
}
\value{
A list object containind

*lambda1_min:* Selected value of lambda1 based on cross validation.
*lambda2_min:* Selected value of lambda1 based on cross validation.
*L_fit:* Estimate of L corresponding to the best fit.
*lambda1_seq:* lambda1 grid used in cross validation.
*lambda2_seq:* lambda2 grid used in cross validation.
}
\description{
Title
}
\examples{
set.seed(11)
require(varband)
n = 100
p = 50
L_true = generateL(p = p, case = "c")$L
X = sample_gen(L = L_true, n = n)
L_cv = smoothcholCV(X, both.lambda = FALSE, n_lambda = 30, pen.type = "fused")


}
